+++ src/routes/datasets/components/DataSamplesList.tsx
@@ -1,9 +1,13 @@
 /** @jsxRuntime classic */
 
 /** @jsx jsx */
+import { IconButton } from '@chakra-ui/button';
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import { css, jsx } from '@emotion/react';
 import styled from '@emotion/styled';
+import { RiFileDownloadLine } from 'react-icons/ri';
+
+import { useAppSelector } from '@/hooks';
 
 import CopyButton from '@/components/CopyButton';
 
@@ -11,6 +15,14 @@ import { Colors, Fonts, Spaces } from '@/assets/theme';
 
 const Container = styled.div`
     position: relative;
+`;
+
+const ActionsContainer = styled.div`
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-end;
+    align-items: center;
+    border-bottom: 1px solid ${Colors.border};
 `;
 
 const copyButtonStyles = css`
@@ -22,8 +34,8 @@ const copyButtonStyles = css`
 const Title = styled.div`
     font-weight: bold;
     padding-bottom: ${Spaces.small};
-    border-bottom: 1px solid ${Colors.border};
     color: ${Colors.content};
+    margin-right: ${Spaces.small};
 `;
 
 const Li = styled.li`
@@ -34,18 +46,45 @@ const Li = styled.li`
 interface DataSamplesListProps {
     title: string;
     keys: string[];
+    sampleType: string;
 }
 
 const DataSamplesList = ({
     title,
     keys,
+    sampleType,
 }: DataSamplesListProps): JSX.Element => {
+    const dataset = useAppSelector((state) => state.datasets.dataset);
+
+    const datasetName = dataset?.name || '';
+    const datasetKey = dataset?.key || '';
+
+    const onDownload = () => {
+        const content = JSON.stringify(keys);
+        const fileName = `${datasetName}_${datasetKey}_${sampleType}_data_samples`;
+
+        const a = document.createElement('a');
+        const file = new Blob([content], { type: 'application/json' });
+        a.href = URL.createObjectURL(file);
+        a.download = fileName;
+        a.click();
+    };
+
     return (
         <Container>
             <Title>{title}</Title>
             <CopyButton css={copyButtonStyles} value={JSON.stringify(keys)}>
                 copy as json array
             </CopyButton>
+            <ActionsContainer>
+                {keys.length > 0 && (
+                    <IconButton
+                        aria-label="download as a json file"
+                        icon={<RiFileDownloadLine />}
+                        onClick={onDownload}
+                    />
+                )}
+            </ActionsContainer>
             <ul>
                 {keys.map((key) => (
                     <Li key={key}>{key}</Li>